% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_probs.R
\name{write_probs}
\alias{write_probs}
\title{Write genotype probabilities to database}
\usage{
write_probs(db, probs, map = NULL, url = "mongodb://localhost",
  quiet = TRUE)
}
\arguments{
\item{db}{Name of database}

\item{probs}{Genotype probabilities, as calculated by \code{\link[qtl2geno]{calc_genoprob}}.}

\item{map}{Optional map of marker positions (a list of vectors of positions).}

\item{url}{URL for mongo server}

\item{quiet}{If FALSE, print some information about progress.}
}
\value{
None. (See Details.)
}
\description{
Write genotype probabilities to database
}
\details{
The genotype probabilities are written to a single Mongo database in a pair of collections.

The data are placed in a series of tables.
\itemize{
\item \code{probs} - the probabilities, with fields \code{marker},
    \code{chr}, \code{marker_index}, \code{probs}, and potentially
    \code{pos}.
\item \code{chr} - chromosome information, with fields \code{chr}, \code{is_x_chr}, \code{geno}
\item \code{meta} - Other information, with fields \code{ind},
    \code{alleles}, \code{crosstype}, \code{alleleprobs}
}
}
\examples{
\dontrun{
library(qtl2geno)
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))
map <- insert_pseudomarkers(iron, map, error_prob=0.002)
pr <- calc_genoprob(iron, map)

write_probs(db="iron_probs", pr, map)
}

}
\seealso{
\code{\link{read_probs}}
}
